+++ src/utils/FileUtils.ts
@@ -117,6 +117,8 @@ export class FileUtils {
 			filename = filename.replace(/\{\{editionId\}\}/g, String(editionIdValue));
 		}
 
+		console.log({ availableData });
+
 		// clean up empty brackets and extra spacing
 		filename = filename
 			.replace(/\(\s*\)/g, "")
@@ -140,12 +142,12 @@ export class FileUtils {
 				.replace(/\]/g, "\\]")
 				// obsidian formatting chars
 				.replace(/\*/g, "\\*")
-				.replace(/\_/g, "\\_")
-				.replace(/\`/g, "\\`")
+				.replace(/_/g, "\\_")
+				.replace(/`/g, "\\`")
 				// headings and tags
-				.replace(/\#/g, "\\#")
+				.replace(/#/g, "\\#")
 				// HTML tags
-				.replace(/\</g, "\\<")
+				.replace(/</g, "\\<")
 		);
 	}